Separated cookbook into sections
Finding recipes is difficult when they're not split into categories. By
separating the recipes into sections, it makes maintenance and reworks
easier in the future (can just modify the section and recompile the
document instead of finding it within a larger document.) In addition it
should increase usability, since separate sections means a more
organized cookbook.